+++ Laboratorio2/Alumno/TareasAlumno.aspx.cs
@@ -7,39 +7,33 @@ using System.Web.UI.WebControls;
 
 namespace Laboratorio2
 {
-    public partial class TareasProfesor : System.Web.UI.Page
+    public partial class TareasAlumno : System.Web.UI.Page
     {
         protected void Page_Load(object sender, EventArgs e)
         {
-            /*
-            if (Session["Tipo"] != null)
-            {
-                if (Session["Tipo"] != "Prof")
-                {
-                    Response.Redirect("~/Inicio.aspx");
-                }
-            }
-            else
-            { //Control de que no venga directamente.
-                Response.Redirect("~/Inicio.aspx");
-            }
-            */
+          
+        }
+
+        protected void SqlDataSource1_Selecting(object sender, SqlDataSourceSelectingEventArgs e)
+        {
 
         }
 
         protected void DropDownList1_SelectedIndexChanged(object sender, EventArgs e)
         {
-            
+
         }
 
         protected void GridView1_SelectedIndexChanged(object sender, EventArgs e)
         {
-            
+            Response.Redirect("~/Alumno/InstanciarTarea.aspx?codigo="+ GridView1.SelectedRow.Cells[1].Text + "&he="+ GridView1.SelectedRow.Cells[3].Text);
         }
 
         protected void Button1_Click(object sender, EventArgs e)
         {
-            Response.Redirect("~/InsertarTarea.aspx");
+            Session.Clear();
+            System.Web.Security.FormsAuthentication.SignOut();
+            Response.Redirect("~/Inicio.aspx");
         }
     }
 }
