+++ Laboratorio2/Alumno/Web.config
@@ -10,13 +10,31 @@
       providerName="System.Data.SqlClient" />
   </connectionStrings>
 	
-  <system.web>
-	<authentication mode="Forms">
-		<forms loginUrl="Inicio.aspx" protection="All"/>
-	</authentication>
-    <compilation debug="true" targetFramework="4.7.2"/>
-    <httpRuntime targetFramework="4.7.2"/>
-  </system.web>
+
+		<location path="InstanciarTarea.aspx">
+		<system.web>
+			<authorization>
+				<allow users="alum" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
+	<location path="TareasAlumno.aspx">
+		<system.web>
+			<authorization>
+				<allow users="alum" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
+	<location path="Alumno.aspx">
+		<system.web>
+			<authorization>
+				<allow users="alum" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
 
 	<system.codedom>
     <compilers>
