+++ Laboratorio2/Profesor/Profesor.aspx
@@ -11,11 +11,11 @@
     <form id="form1" runat="server">
         <div itemtype="TareasProfesor.aspx" style="border-style: solid; border-width: thin; width: 483px; height: 299px;">
             &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-            <asp:HyperLink ID="HyperLink1" runat="server" NavigateUrl="~/TareasProfesor.aspx">Tareas</asp:HyperLink>
+            <asp:HyperLink ID="HyperLink1" runat="server" NavigateUrl="~/Profesor/TareasProfesor.aspx">Tareas</asp:HyperLink>
             <br />
-            <asp:HyperLink ID="HyperLink2" runat="server" NavigateUrl="~/ProfesorImportar.aspx">IMPORTAR .XML</asp:HyperLink>
+            <asp:HyperLink ID="HyperLink2" runat="server" NavigateUrl="~/Profesor/ProfesorImportar.aspx">IMPORTAR .XML</asp:HyperLink>
             <br />
-            <asp:HyperLink ID="HyperLink3" runat="server" NavigateUrl="~/ProfesorExportar.aspx">Exportar .XML</asp:HyperLink>
+            <asp:HyperLink ID="HyperLink3" runat="server" NavigateUrl="~/Profesor/ProfesorExportar.aspx">Exportar .XML</asp:HyperLink>
         </div>
     </form>
 </body>
