+++ Laboratorio2/Profesor/ProfesorExportar.aspx.cs
@@ -14,19 +14,7 @@ namespace Laboratorio2
     {
         protected void Page_Load(object sender, EventArgs e)
         {
-            /*
-            if (Session["Tipo"] != null)
-            {
-                if (Session["Tipo"] != "Prof")
-                {
-                    Response.Redirect("~/Inicio.aspx");
-                }
-            }
-            else
-            { //Control de que no venga directamente.
-                Response.Redirect("~/Inicio.aspx");
-            }
-            */
+
         }
 
         protected void DropDownList1_SelectedIndexChanged(object sender, EventArgs e)
@@ -90,5 +78,11 @@ namespace Laboratorio2
         {
 
         }
+
+        protected void DropDownList1_DataBound(object sender, EventArgs e)
+        {
+
+        }
+
     }
 }
