+++ Laboratorio2/Profesor/Web.config
@@ -10,13 +10,47 @@
       providerName="System.Data.SqlClient" />
   </connectionStrings>
 	
-  <system.web>
-	<authentication mode="Forms">
-		<forms loginUrl="Inicio.aspx" protection="All"/>
-	</authentication>
-    <compilation debug="true" targetFramework="4.7.2"/>
-    <httpRuntime targetFramework="4.7.2"/>
-  </system.web>
+
+	<location path="ProfesorExportar.aspx">
+		<system.web>
+			<authorization>
+				<allow users="vadillo" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
+	<location path="ProfesorImportar.aspx">
+		<system.web>
+			<authorization>
+				<allow users="vadillo" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
+	<location path="Profesor.aspx">
+		<system.web>
+			<authorization>
+				<allow users="vadillo,prof" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
+	<location path="InsertarTarea.aspx">
+		<system.web>
+			<authorization>
+				<allow users="vadillo,prof" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
+		<location path="TareasProfesor.aspx">
+		<system.web>
+			<authorization>
+				<allow users="vadillo,prof" />
+				<deny users="*" />
+			</authorization>
+		</system.web>
+	</location>
 
 	<system.codedom>
     <compilers>
